feat: Create materialized view for FocusClean table
This commit adds a new materialized view called FocusClean on the table Focus. The view is created in the adxTableSchema.kql file and adxViews.kql file. The purpose of this view is to summarize the data in the Focus table based on various columns such as BillingAccountId, ChargePeriodStart, CommitmentDiscountId, Region, ResourceId, SkuPriceId, SubAccountId, x_AccountOwnerId, x_CostCenter, x_InvoiceSectionId, x_SkuDetails, x_SkuMeterId, and x_SkuOfferId. The view is created with the option to backfill existing data.