Update Deploy-FinOpsHub.ps1
when expressing a [switch] as a boolean in a hashtable you need to express the .IsPresent property.

enableInfrastructureEncryption was being parsed as an object rather than a boolean causing the error:

Error: Code=InvalidTemplate; Message=Deployment template validation failed: 'The provided value for the template parameter 'enableInfrastructureEncryption' is not valid. Expected a value of type 'Boolean', but received a value of type 'Object'.